Final version for the migrated package
